refactor(translator): rename gemini translator module to antigravity
- Rename packages/translator/src/gemini.ts to antigravity.ts
- Rename tests/gemini.test.ts to tests/antigravity.test.ts
- Update export barrel in packages/translator/src/index.ts
