+++ packages/translator/src/index.ts
@@ -1,7 +1,7 @@
 export * from "./adapter.js";
 export * from "./anthropic.js";
+export * from "./antigravity.js";
 export * from "./commandcode.js";
-export * from "./gemini.js";
 export * from "./responses.js";
 export * from "./usage.js";
 export * from "./tokenSaver.js";
